refactor: 💡 move navigation code to MainNavHost
